+++ hata/discord/auto_moderation/tests/test__AutoModerationRuleTriggerType.py
@@ -5,7 +5,7 @@ from .. import AutoModerationRuleTriggerType, AutoModerationRuleTriggerMetadata
 
 def test__AutoModerationRuleTriggerType__name():
     """
-    Tests whether ``AutoModerationRuleTriggerType` instance names are all strings.
+    Tests whether ``AutoModerationRuleTriggerType`` instance names are all strings.
     """
     for instance in AutoModerationRuleTriggerType.INSTANCES.values():
         vampytest.assert_instance(instance.name, str)
@@ -13,7 +13,7 @@ def test__AutoModerationRuleTriggerType__name():
 
 def test__AutoModerationRuleTriggerType__value():
     """
-    Tests whether ``AutoModerationRuleTriggerType` instance values are all the expected value type.
+    Tests whether ``AutoModerationRuleTriggerType`` instance values are all the expected value type.
     """
     for instance in AutoModerationRuleTriggerType.INSTANCES.values():
         vampytest.assert_instance(instance.value, AutoModerationRuleTriggerType.VALUE_TYPE)
@@ -21,7 +21,7 @@ def test__AutoModerationRuleTriggerType__value():
 
 def test__AutoModerationRuleTriggerType__max_per_guild():
     """
-    Tests whether ``AutoModerationRuleTriggerType` instance `.max_per_guild`-s are all ints.
+    Tests whether ``AutoModerationRuleTriggerType`` instance `.max_per_guild`-s are all ints.
     """
     for instance in AutoModerationRuleTriggerType.INSTANCES.values():
         vampytest.assert_instance(instance.max_per_guild, int)
@@ -29,7 +29,7 @@ def test__AutoModerationRuleTriggerType__max_per_guild():
 
 def test__AutoModerationRuleTriggerType__metadata_type():
     """
-    Tests whether ``AutoModerationRuleTriggerType` instance `.metadata_type`-s are all
+    Tests whether ``AutoModerationRuleTriggerType`` instance `.metadata_type`-s are all
     ``AutoModerationRuleTriggerMetadata``-s or `None`-s.
     """
     for instance in AutoModerationRuleTriggerType.INSTANCES.values():